+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 
 New functionality added in a backwards-compatible manner:
 * [OLMIS-7308](https://openlmis.atlassian.net/browse/OLMIS-7308): Added extension module for StockEvent postprocess.
+* [OLMIS-7305](https://openlmis.atlassian.net/browse/OLMIS-7305): Added sending stockEvent object to one-network-integration-service.